feat: add offline mode in DetailScreen

+++ app/src/main/java/com/pascal/ecommercecompose/data/local/dao/ProductDao.kt
@@ -10,7 +10,7 @@ import com.pascal.ecommercecompose.data.local.entity.ProductEntity
 @Dao
 interface ProductDao {
     @Query("SELECT * FROM Product WHERE id = :id")
-    suspend fun getProductById(id: Long): ProductEntity?
+    suspend fun getProductById(id: Int): ProductEntity?
 
     @Query("SELECT * FROM Product")
     suspend fun getAllProducts(): List<ProductEntity>
